solved the generaticstaticparams error
