+++ src/components/ApplyForm/ApplyForm.js
@@ -57,18 +57,11 @@ const ApplyForm = () => {
                 max: 10,
               },
               {
-                fieldName: "Branch",
-                inputType: "text",
+                fieldName: "Year",
+                inputType: "dropdown",
                 datatype: "text",
-                placeholder: "Enter your branch (e.g., Computer Science, IT)",
-                max: 100,
-              },
-              {
-                fieldName: "Class",
-                inputType: "text",
-                datatype: "text",
-                placeholder: "Enter your class (e.g., SE1, TE1)",
-                max: 50,
+                options: ["SE", "TE", "BE"],
+                placeholder: "Select your Engineering year",
               },
             ],
           },
@@ -76,25 +69,25 @@ const ApplyForm = () => {
             title: "Technical Info",
             fields: [
               {
-                fieldName: "Interest",
+                fieldName: "Preferred roles",
                 inputType: "dropdown",
                 datatype: "text",
-                options: ["Frontend", "Backend", "Design"],
+                options: ["Frontend", "Backend", "Design", "Cloud", "FullStack"],
                 placeholder: "Select your area of interest",
               },
               {
-                fieldName: "Programming Languages Known",
+                fieldName: "Skills",
                 inputType: "textarea",
                 datatype: "text",
-                placeholder: "List the programming languages you know (e.g., JavaScript, Python, Java)",
+                placeholder: "List the programming languages you know (e.g., JavaScript, React.js, Node.js)",
                 max: 1000,
               },
               {
-                fieldName: "Years of Experience",
-                inputType: "text",
-                datatype: "number",
-                placeholder: "Enter your years of experience (if any)",
-                max: 50,
+                fieldName: "Innovative Idea",
+                inputType: "textarea",
+                datatype: "text",
+                placeholder: "Any innovative idea you possess to add to the website",
+                max: 10000,
               },
             ],
           },
@@ -102,18 +95,18 @@ const ApplyForm = () => {
             title: "Portfolio & Resume",
             fields: [
               {
-                fieldName: "Portfolio URL",
-                inputType: "text",
-                datatype: "url",
-                placeholder: "Enter the URL to your portfolio or GitHub repository",
-                max: 200,
+                fieldName: "Projects Worked On",
+                inputType: "textarea",
+                datatype: "text",
+                placeholder: "Projects they've worked on (Brief description + tech stack used), GitHub/project link",
+                max: 5000,
               },
               {
-                fieldName: "Resume URL",
-                inputType: "text",
-                datatype: "url",
-                placeholder: "Enter the URL to your resume",
-                max: 200,
+                fieldName: "Social Profiles",
+                inputType: "textarea",
+                datatype: "text",
+                placeholder: "GitHub profile/Portfolio website link",
+                max: 1000,
               },
             ],
           },
