+++ src/components/Events/EventSlider.js
@@ -2,17 +2,18 @@ import React from 'react'
 import EventsList from './EventsList'
 import EventCarousell from './EventCarousell'
 
-const EventSlider = ({events}) => {
-    const pastEvents = events.past;
-    const ongoingEvents = events.ongoing;
-    const upcomingEvents = events.upcoming;
+const EventSlider = ({ events }) => {
+  const pastEvents = events.past;
+  const ongoingEvents = events.ongoing;
+  const upcomingEvents = events.upcoming;
 
   return (
     <div>
-        <EventCarousell ongoingEvents={ongoingEvents} />
+      <EventCarousell ongoingEvents={ongoingEvents} />
       <div className=" mx-auto py-8">
-        <EventsList title="Past Events" events={pastEvents} />
         <EventsList title="Upcoming Events" events={upcomingEvents} />
+        <EventsList title="Past Events" events={pastEvents} />
+
       </div>
     </div>
   )
